requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
pyTelegramBotAPI==4.5.1
python-dotenv==0.20.0
requests==2.27.1
schedule==1.1.0
urllib3==1.26.9
